fix: removed circular dependency, moved jira config to types/config.ts
