+++ agent-packages/packages/jira/src/types/config.ts
@@ -1,0 +1,23 @@
+import { BaseConfig } from '@clearfeed-ai/quix-common-agent';
+
+export type JiraAuth =
+  | {
+      username: string;
+      password: string;
+    }
+  | {
+      bearerToken: string;
+    }
+  | {
+      sharedSecret: string;
+      atlassianConnectAppKey: string;
+    };
+
+export interface JiraConfig extends BaseConfig {
+  host: string;
+  defaultConfig?: {
+    projectKey?: string;
+  };
+  auth: JiraAuth;
+  apiHost?: string;
+}
